implement getFolderItems and getItemInfo in configurated-citiation/storage-addon child classes
