+++ app/models/configured-citation-addon.ts
@@ -1,6 +1,9 @@
 import { AsyncBelongsTo, belongsTo } from '@ember-data/model';
 
 import ResourceReferenceModel from 'ember-osf-web/models/resource-reference';
+import { task } from 'ember-concurrency';
+import { waitFor } from '@ember/test-waiters';
+import { ConnectedCitationOperationNames, OperationKwargs } from 'ember-osf-web/models/addon-operation-invocation';
 import AuthorizedCitationAccountModel from './authorized-citation-account';
 import ExternalCitationServiceModel from './external-citation-service';
 import ConfiguredAddonModel from './configured-addon';
@@ -17,6 +20,31 @@ export default class ConfiguredCitationAddonModel extends ConfiguredAddonModel {
 
     get externalServiceId() {
         return (this as ConfiguredCitationAddonModel).belongsTo('externalCitationService').id();
+    }
+
+    @task
+    @waitFor
+    async getFolderItems(this: ConfiguredAddonModel, kwargs?: OperationKwargs) {
+        const operationKwargs = kwargs || {};
+        const operationName = operationKwargs.itemId ? ConnectedCitationOperationNames.ListCollectionItems :
+            ConnectedCitationOperationNames.ListRootCollections;
+        const newInvocation = this.store.createRecord('addon-operation-invocation', {
+            operationName,
+            operationKwargs,
+            thruAddon: this,
+        });
+        return await newInvocation.save();
+    }
+
+    @task
+    @waitFor
+    async getItemInfo(this: ConfiguredAddonModel, itemId: string) {
+        const newInvocation = this.store.createRecord('addon-operation-invocation', {
+            operationName: ConnectedCitationOperationNames.GetItemInfo,
+            operationKwargs: { itemId },
+            thruAddon: this,
+        });
+        return await newInvocation.save();
     }
 }
