+++ app/models/configured-storage-addon.ts
@@ -1,4 +1,7 @@
 import { AsyncBelongsTo, attr, belongsTo } from '@ember-data/model';
+import { waitFor } from '@ember/test-waiters';
+import { task } from 'ember-concurrency';
+import { ConnectedStorageOperationNames, OperationKwargs } from 'ember-osf-web/models/addon-operation-invocation';
 import ResourceReferenceModel from 'ember-osf-web/models/resource-reference';
 
 import AuthorizedStorageAccountModel from './authorized-storage-account';
@@ -20,6 +23,31 @@ export default class ConfiguredStorageAddonModel extends ConfiguredAddonModel {
 
     get externalServiceId() {
         return (this as ConfiguredStorageAddonModel).belongsTo('externalStorageService').id();
+    }
+
+    @task
+    @waitFor
+    async getFolderItems(this: ConfiguredAddonModel, kwargs?: OperationKwargs) {
+        const operationKwargs = kwargs || {};
+        const operationName = operationKwargs.itemId ? ConnectedStorageOperationNames.ListChildItems :
+            ConnectedStorageOperationNames.ListRootItems;
+        const newInvocation = this.store.createRecord('addon-operation-invocation', {
+            operationName,
+            operationKwargs,
+            thruAddon: this,
+        });
+        return await newInvocation.save();
+    }
+
+    @task
+    @waitFor
+    async getItemInfo(this: ConfiguredAddonModel, itemId: string) {
+        const newInvocation = this.store.createRecord('addon-operation-invocation', {
+            operationName: ConnectedStorageOperationNames.GetItemInfo,
+            operationKwargs: { itemId },
+            thruAddon: this,
+        });
+        return await newInvocation.save();
     }
 }
